requirements.txt
flask==2.0.1
flask-socketio==5.1.1
python-dotenv==0.19.0
click==8.0.1
watchdog==2.1.5
psutil==5.8.0
netifaces==0.11.0
speedtest-cli==2.1.3
python-nmap==0.7.1
scapy==2.4.5
pyroute2==0.6.5
pyaml==21.8.3
jsonschema==3.2.0
bcrypt==3.2.0
python-crontab==2.5.1
PyJWT==2.1.0
requests==2.26.0
paho-mqtt==1.5.1
